S302T700-211 #time 30m implement deletion of image ?implement

+++ backend/src/main/java/com/seng302/wasteless/controller/ImageController.java
@@ -125,9 +125,20 @@ public class ImageController {
 
         ProductImage image = productImageService.findImageById(imageId);
         Product product = productService.findProductById(productId);
-        productImageService.deleteImage(image);
-        productService.removeImageFromProduct(product, imageId);
+
+        if (product==null){
+            logger.warn("Cannot delete productImage. Product is null");
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Product no longer exists");
+        }
+        if (image==null){
+            logger.warn("Cannot delete productImage. image: is null", image, product);
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body("Image no longer exists");
+        }
+
+        productService.removeImageFromProduct(product, image);
         productService.updateProduct(product);
+        productImageService.deleteImage(image);
+        productImageService.deleteImageFile(image);
         return ResponseEntity.status(HttpStatus.OK).build();
     }
